change -a to simple flag, no argument.

+++ WorkFlowASTER_onestrip.sh
@@ -20,9 +20,10 @@ CorThr=0.7
 SzW=5
 water_mask=false
 do_ply=false
+do_angle=false
 NoCorDEM=false
 
-while getopts "s:z:c:q:wnf:t:yh" opt; do
+while getopts "s:z:c:q:wnf:t:y:ah" opt; do
   case $opt in
     h)
       echo "Run the second step in the MMASTER processing chain."
@@ -31,11 +32,12 @@ while getopts "s:z:c:q:wnf:t:yh" opt; do
       echo "    -z UTMZONE  : UTM Zone of area of interest. Takes form 'NN +north(south)'"
       echo "    -c CorThr   : Correlation Threshold for estimates of Z min and max (optional, default : 0.7)"
       echo "    -q SzW      : Size of the correlation window in the last step (optional, default : 4, mean 9*9)"
-      echo "    -w mask     : Mask large water areas."
+      echo "    -w mask     : Name of shapefile to skip masked areas (usually water, this is optional, default : none)."
       echo "    -n NoCorDEM : Compute DEM with the uncorrected 3B image (computing with correction as well)"
       echo "    -f ZOOMF    : Run with different final resolution   (optional; default: 1)"
       echo "    -t RESTERR  : Run with different terrain resolution (optional; default: 30)"
       echo "    -y do_ply   : Write point cloud (DEM drapped with ortho in ply)"
+      echo "    -a do_angle : Compute track angle along orbit"
       echo "    -h          : displays this message and exits."
       echo " "
       exit 0
@@ -43,6 +45,9 @@ while getopts "s:z:c:q:wnf:t:yh" opt; do
     n)
       NoCorDEM=$OPTARG
       ;;
+    a)
+      do_angle=true
+      ;;  
     y)
       do_ply=$OPTARG
       ;;    
@@ -63,7 +68,8 @@ while getopts "s:z:c:q:wnf:t:yh" opt; do
       echo "SzW set to $SzW"
       ;;
     w)
-      water_mask=$OPTARG
+e     echo "Water mask selected: " $OPTARG
+	  nameWaterMask=$OPTARG
       ;;
     f)
       ZoomF=$OPTARG
@@ -169,9 +175,8 @@ mv $name$Bt ImOrig/$name$Bt
 mv $name$Bcor $name$Bt
 
 # if we're using a water mask, we run that here.
-if [ "$water_mask" = true ]; then #check variable name!
-    cp $MMSCRIPTS_HOME/WorkFlow_WaterMask.sh .
-    bash WorkFlow_WaterMask.sh $name $UTM
+if [ "$nameWaterMask" != false ]; then #check variable name!
+    WorkFlow_WaterMask.sh $name "$UTM" $nameWaterMask
 fi
 
 # Correlation with corrected image
@@ -187,33 +192,51 @@ fi
 cd MEC-Malt
 mv Correl_STD-MALT_Num_8.tif Correl_STD-MALT_Num_8_FullRes.tif
 cp Z_Num9_DeZoom1_STD-MALT.tfw Correl_STD-MALT_Num_8_FullRes.tfw
-gdal_translate -tr $RESTERR $RESTERR -r cubicspline -a_srs "+proj=utm +zone=$UTM +datum=WGS84 +units=m +no_defs" Correl_STD-MALT_Num_8_FullRes.tif Correl_STD-MALT_Num_8.tif
+gdal_translate -tr $RESTERR $RESTERR -a_srs "+proj=utm +zone=$UTM +datum=WGS84 +units=m +no_defs" Correl_STD-MALT_Num_8_FullRes.tif Correl_STD-MALT_Num_8.tif
+
 mv AutoMask_STD-MALT_Num_8.tif AutoMask_STD-MALT_Num_8_FullRes.tif
 cp Z_Num9_DeZoom1_STD-MALT.tfw AutoMask_STD-MALT_Num_8_FullRes.tfw
 gdal_translate -tr $RESTERR $RESTERR -r cubicspline -a_srs "+proj=utm +zone=$UTM +datum=WGS84 +units=m +no_defs" AutoMask_STD-MALT_Num_8_FullRes.tif AutoMask_STD-MALT_Num_8.tif
+
+if [ -f Z_Num9_DeZoom1_STD-MALT_Tile_0_0.tif ]; then
+	mosaic_micmac_tiles.py -filename 'Z_Num9_DeZoom1_STD-MALT' 
+fi
 mv Z_Num9_DeZoom1_STD-MALT.tif Z_Num9_DeZoom1_STD-MALT_FullRes.tif
 mv Z_Num9_DeZoom1_STD-MALT.tfw Z_Num9_DeZoom1_STD-MALT_FullRes.tfw
 mv Z_Num9_DeZoom1_STD-MALT.xml Z_Num9_DeZoom1_STD-MALT_FullRes.xml
+
 gdal_translate -tr $RESTERR $RESTERR -r cubicspline -a_srs "+proj=utm +zone=$UTM +datum=WGS84 +units=m +no_defs" -co TFW=YES Z_Num9_DeZoom1_STD-MALT_FullRes.tif Z_Num9_DeZoom1_STD-MALT.tif
 cd ..
 
-# computing orbit angles on DEM
-mm3d SateLib ASTERProjAngle MEC-Malt/Z_Num9_DeZoom1_STD-MALT MEC-Malt/AutoMask_STD-MALT_Num_8.tif $name$N
-cp MEC-Malt/Z_Num9_DeZoom1_STD-MALT.tfw TrackAngleMap_nonGT.tfw
-mv TrackAngleMap.tif TrackAngleMap_nonGT.tif
-gdal_translate -a_srs "+proj=utm +zone=$UTM +datum=WGS84 +units=m +no_defs" -a_nodata 0 TrackAngleMap_nonGT.tif TrackAngleMap.tif
-rm TrackAngleMap_nonGT*
-
+if [ "$do_angle" = true ]; then
+	# computing orbit angles on DEM
+	mm3d SateLib ASTERProjAngle MEC-Malt/Z_Num9_DeZoom1_STD-MALT MEC-Malt/AutoMask_STD-MALT_Num_8.tif $name$N
+	if [ -f TrackAngleMap_3N_Tile_0_0.tif ]; then
+		mosaic_micmac_tiles.py -filename 'TrackAngleMap_3N'
+	fi
+	cp MEC-Malt/Z_Num9_DeZoom1_STD-MALT.tfw TrackAngleMap_nonGT.tfw
+	mv TrackAngleMap.tif TrackAngleMap_nonGT.tif
+	gdal_translate -a_srs "+proj=utm +zone=$UTM +datum=WGS84 +units=m +no_defs" -a_nodata 0 TrackAngleMap_nonGT.tif TrackAngleMap_3N.tif
+	rm TrackAngleMap_nonGT*
+	mm3d SateLib ASTERProjAngle MEC-Malt/Z_Num9_DeZoom1_STD-MALT MEC-Malt/AutoMask_STD-MALT_Num_8.tif $name$B
+	if [ -f TrackAngleMap_3B_Tile_0_0.tif ]; then
+		mosaic_micmac_tiles.py -filename 'TrackAngleMap_3B'
+	fi
+	cp MEC-Malt/Z_Num9_DeZoom1_STD-MALT.tfw TrackAngleMap_nonGT.tfw
+	mv TrackAngleMap.tif TrackAngleMap_nonGT.tif
+	gdal_translate -a_srs "+proj=utm +zone=$UTM +datum=WGS84 +units=m +no_defs" -a_nodata 0 TrackAngleMap_nonGT.tif TrackAngleMap_3B.tif
+	rm TrackAngleMap_nonGT*
+fi
 
 
 
 cd Ortho-MEC-Malt
 # if there are no tiles, we have nothing to do.
+# not sure if we want to hard-code that the tiles will always be Nx1?
 if [ -f Orthophotomosaic_Tile_0_0.tif ]; then
-    montage Orthophotomosaic_T*.tif -mode concatenate -tile 1x  Orthophotomosaic_FullRes.tif
-else
-    cp Orthophotomosaic.tif Orthophotomosaic_FullRes.tif
+	mosaic_micmac_tiles.py -filename 'Orthophotomosaic'
 fi
+mv Orthophotomosaic.tif Orthophotomosaic_FullRes.tif
 mv Orthophotomosaic.tfw Orthophotomosaic_FullRes.tfw
-gdal_translate -tr 15 15 -r cubicspline -a_srs "+proj=utm +zone=$UTM +datum=WGS84 +units=m +no_defs" Orthophotomosaic_FullRes.tif Orthophotomosaic.tif
+gdal_translate -tr 15 15 -r bilinear -a_srs "+proj=utm +zone=$UTM +datum=WGS84 +units=m +no_defs" Orthophotomosaic_FullRes.tif Orthophotomosaic.tif
 cd ..
